Display versions and features of dedicated plans of providers
Displaying a provider now prints the supported versions and options each
plan has available. It also sorts the plans based on price instead of
having them mixed up.

Shared plans do not display their options and versions because those are
bound to a cluster, not to the plan. We might need a different display
approach to display it.